+++ tests/conftest.py
@@ -10,6 +10,12 @@ DSM_OWNER = '0x3e40D73EB977Dc6a537aF587D48316feE66E9C8c'
 # https://goerli.etherscan.io/address/0xe57025E250275cA56f92d76660DEcfc490C7E79A#readContract#F12
 # DSM_OWNER = '0xa5F1d7D49F581136Cf6e58B32cBE9a2039C48bA1'
 
+COUNCIL_ADDRESS_1 = '0x70997970C51812dc3A010C7d01b50e0d17dc79C8'
+COUNCIL_PK_1 = '0x59c6995e998f97a5a0044966f0945389dc9e86dae88c7a8412f4603b6b78690d'
+
+COUNCIL_ADDRESS_2 = '0x3C44CdDdB6a900fa2b585dd299e03d12FA4293BC'
+COUNCIL_PK_2 = '0x5de4111afa1a4b94908f83103eb1f1706367c2e68ca870fc3fb9a804cdab365a'
+
 
 @pytest.fixture
 def block_data():
@@ -43,3 +49,14 @@ def ping_message():
         'guardianAddress': '0x3dc4cF780F2599B528F37dedB34449Fb65Ef7d4A',
         'app': {'version': '1.1.1', 'name': 'lido-council-daemon'},
     }
+
+
+@pytest.fixture
+def add_accounts_to_guardian(web3_lido_integration, set_integration_account):
+    web3_lido_integration.provider.make_request('anvil_impersonateAccount', [DSM_OWNER])
+    web3_lido_integration.provider.make_request('anvil_setBalance', [DSM_OWNER, '0x500000000000000000000000'])
+
+    web3_lido_integration.lido.deposit_security_module.functions.addGuardian(COUNCIL_ADDRESS_1, 2).transact({'from': DSM_OWNER})
+    web3_lido_integration.lido.deposit_security_module.functions.addGuardian(COUNCIL_ADDRESS_2, 2).transact({'from': DSM_OWNER})
+
+    yield web3_lido_integration
